Enhanced Styles and detailt it

+++ src/components/Iterations.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
 import axios from "axios";
 import Global from '../Global';
-import { Link } from "react-router-dom";
+import { Link, Redirect } from "react-router-dom";
+import swal from 'sweetalert';
 
 export default class Iterations extends Component {
     url = Global.url;
@@ -24,25 +25,42 @@ export default class Iterations extends Component {
             });
     }
     deleteIteration = (id) => {
-        var it = {
-            "iterationId": id
-        };
-        axios.delete(this.url + "iteration/delete", 
-            {headers: {
-                "Content-Type": "application/json"
-            },
-            data: it
-            }
-          )
-          .then(res =>{
-                console.log(res.data);
-                this.setState({
-                    iterations: res.data,
-                    status: 'success'
-                })
-            });
+        swal({
+            title: "Estas Seguro?",
+            text: "Si eliminas este registro no podra ser recuperado",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+                var it = {
+                    "iterationId": id
+                };
+                axios.delete(this.url + "iteration/delete", 
+                    {headers: {
+                        "Content-Type": "application/json"
+                    },
+                    data: it
+                    }
+                )
+                .then(res =>{
+                    console.log(res.data);
+                    this.setState({
+                        iterations: res.data,
+                        status: 'success'
+                    });
+                    swal("Iteracion Eliminada Correctamente", {
+                      icon: "delete",
+                    });
+                });
+            } 
+          });
     }
     render(){
+        if(this.state.status === 'delete'){
+            return <Redirect to="/"></Redirect>;
+        }
         var listIteration = this.state.iterations.map((it) =>{
             return(
                 <tr key={it._id}>
@@ -50,21 +68,21 @@ export default class Iterations extends Component {
                     <td>{it.description}</td>
                     <td>{it.betaVerificator}</td>
                     <td>
-                        <Link to={'/iteration/view/' + it._id}>Detalle</Link>
-                        <Link to={'/iteration/edit/' + it._id}>Editar</Link>
+                        <Link to={'/iteration/view/' + it._id} className="btn btn-info" type="button">Detalle</Link>
+                        <Link to={'/iteration/edit/' + it._id} className="btn btn-warning" type="button">Editar</Link>
                         <button onClick={
                             () => {
                                 this.deleteIteration(it._id);
                             }
-                        }>Eliminar</button>
+                        } className="btn btn-danger" type="button">Eliminar</button>
                     </td>
                 </tr>
             );
         });
         return(
             <div id="iterations">
-                <Link to={'/iteration/create'}>Crear Iteración</Link>
-                <table border="1">
+                <Link to={'/iteration/create'} className="btn btn-primary" type="button">Crear Iteración</Link>
+                <table  className="table">
                     <thead>
                         <tr>
                             <th>Nombre</th>
